+++ docs/release-notes.md
@@ -20,6 +20,9 @@
   * Fixed mods which implement `IAssetLoader` directly not being allowed to load files due to incorrect conflict detection.
   * Fixed SMAPI blocking reflection access to vanilla members on overridden types.
 
+* For SMAPI developers:
+  * Added the SMAPI installer version and platform to the window title to simplify troubleshooting.
+
 ## 2.0
 ### Release highlights
 * **Mod update checks**